src/commands/entrypoint.sh: Make sure passt can write its PID file
When running under Docker or rootful Podman, passt will realize that it
is running as *actual* root and will switch to being user nobody, but
this will make it fail to create its PID file because its directory was
created by libvirt and is thus owned by root:root.

We get around that by providing a wrapper script around passt that first
gives "others" write access to the directory.

Signed-off-by: Alberto Faria <afaria@redhat.com>