first_boot: Add context for genisoimage
I got a bare
```
Error: crun-vm: failed to load cloud-init config: No such file or directory (os error 2): OCI runtime attempted to invoke a command that was not found
```

Because my host was missing `genisoimage`.  This turns that into
```
Error: crun-vm: failed to load cloud-init config: Invoking genisoimage: No such file or directory (os error 2): OCI runtime attempted to invoke a command that was not found
```

which is a bit more useful.

Signed-off-by: Colin Walters <walters@verbum.org>