docs: various tweaks from a first-time reader
In step 1, since editing /etc is not desirable for rootless Podman,
mention the alternative location that will have the same effect.

In step 2, there's no need for the disclaimer paragraph about Docker
lacking --latest in an example that already can't use Docker because
of lack of --rootfs a couple paragraphs above.  And even for Podman,
make it clear that the initial example is limited, in that the qcow2
image has no ssh keys, root password, or alternative users installed
yet.

Signed-off-by: Eric Blake <eblake@redhat.com>